+++ lua/cargdev/plugins/yaml.lua
@@ -13,7 +13,6 @@ return {
   dependencies = {
     "nvim-treesitter/nvim-treesitter",
     "folke/snacks.nvim", -- optional
-    "nvim-telescope/telescope.nvim", -- optional
     "ibhagwan/fzf-lua", -- optional
   },
 }